+++ traffic_portal/app/src/common/modules/form/deliveryServiceSslKeys/generate/FormGenerateDeliveryServiceSslKeysController.js
@@ -34,7 +34,25 @@ var FormGenerateDeliveryServiceSslKeysController = function(deliveryService, ssl
 		return sslRequest;
 	};
 
-	$scope.useLetsEncrypt = false;
+	var getAcmeProviders = function() {
+		deliveryServiceSslKeysService.getAcmeProviders()
+			.then(function(result) {
+				$scope.acmeProviders = result;
+				if (!$scope.acmeProviders.includes('Lets Encrypt')) {
+					$scope.acmeProviders.push('Lets Encrypt');
+				}
+			});
+	};
+
+	$scope.loadAcmeProviders = function() {
+		if ($scope.useAcme) {
+			getAcmeProviders();
+		}
+	};
+
+	$scope.useAcme = false;
+	$scope.acmeProviders = [];
+	$scope.acmeProvider = "";
 	$scope.hasError = formUtils.hasError;
 	$scope.hasPropertyError = formUtils.hasPropertyError;
 	$scope.navigateToPath = locationUtils.navigateToPath;
@@ -318,7 +336,7 @@ var FormGenerateDeliveryServiceSslKeysController = function(deliveryService, ssl
         });
     };
 
-    $scope.confirmGenerateLetsEncrypt = function(sslRequest) {
+    $scope.confirmGenerateAcme = function(sslRequest) {
         var params = {
             title: 'Generate New SSL Keys Using Let\'s Encrypt for Delivery Service: ' + deliveryService.xmlId,
             message: ' (replacing any previous keys)'
@@ -334,7 +352,8 @@ var FormGenerateDeliveryServiceSslKeysController = function(deliveryService, ssl
             }
         });
         modalInstance.result.then(function() {
-            deliveryServiceSslKeysService.generateSslKeysWithLetsEncrypt(deliveryService, sslKeys, sslRequest).then(
+			sslKeys.authType = $scope.acmeProvider;
+            deliveryServiceSslKeysService.generateSslKeysWithAcme(deliveryService, sslKeys, sslRequest, $scope.acmeProvider).then(
                 function() {
                     locationUtils.navigateToPath('/delivery-services/' + deliveryService.id + '/ssl-keys');
                 });
