+++ traffic_portal/app/src/scripts/config.js
@@ -23,6 +23,6 @@
 
 angular.module('config', [])
 
-.constant('ENV', { api: { root:'/api/4.0/', legacy: '/api/1.5/' } })
+.constant('ENV', { api: { root:'/api/4.1/', legacy: '/api/1.5/' } })
 
 ;
